+++ Content/minimal/src/Client/fable_modules/Fable.Promise.2.2.2/PromiseImpl.fs
@@ -1,0 +1,4 @@
+[<AutoOpen>]
+module PromiseImpl
+
+let promise = Promise.PromiseBuilder()