+++ Content/minimal/src/Client/fable_modules/fable-library.4.9.0/lib/big.d.ts
@@ -1,0 +1,5 @@
+/**
+ * @type object
+ */
+export const Big: object;
+export default Big;